+++ src/ErrorOr/EmptyErrors.cs
@@ -1,6 +1,6 @@
-﻿namespace ErrorOr;
+namespace ErrorOr;
 
 internal static class EmptyErrors
 {
-    public static List<Error> Instance { get; } = [];
+    public static ReadOnlyCollection<Error> Instance { get; } = new ReadOnlyCollection<Error>([]);
 }